Added initial routing, moved some files
