+++ soos-client/package.json
@@ -14,6 +14,7 @@
     "axios": "^1.2.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.14.2",
     "socket.io-client": "^4.6.1",
     "soos-gamelogic": "file:../soos-gamelogic"
   },
