+++ soos-client/src/routes/Entry.tsx
@@ -1,0 +1,5 @@
+export function Entry() {
+  return (
+    <div>Entry page</div>
+  )
+}
