+++ soos-client/src/routes/Lobby.tsx
@@ -1,0 +1,5 @@
+export function Lobby() {
+  return (
+    <div>Lobby page</div>
+  )
+}
