+++ soos-client/src/scss/App.scss
@@ -1,4 +1,4 @@
-@import './scss/variables.scss';
+@import './variables.scss';
 
 * {
     box-sizing: border-box;
